Create shared library for draft mode
